+++ backend/jest.config.ts
@@ -3,14 +3,11 @@ import type { Config } from 'jest';
 const config: Config = {
   preset: 'ts-jest',
   testEnvironment: 'node',
-  testMatch: [
-    '**/?(*.)+(test).[jt]s?(x)', // allow tests next to source files
-    '**/tests/**/*.test.ts', // keep support for central tests folder
-  ],
+  testMatch: ['**/?(*.)+(test).[jt]s?(x)', '**/tests/**/*.test.ts'],
   moduleNameMapper: {
     '^@/(.*)$': '<rootDir>/src/$1',
     '^@controller/(.*)$': '<rootDir>/src/controller/$1',
-    '^@middlewares/(.*)$': '<rootDir>/src/middlewares/$1',
+    '^@middlewares$': '<rootDir>/src/middlewares/index.ts',
     '^@routes/(.*)$': '<rootDir>/src/routes/$1',
     '^@types/(.*)$': '<rootDir>/src/types/$1',
     '^@prismaClient/(.*)$': '<rootDir>/prisma/$1',
