+++ backend/src/controller/admin/index.ts
@@ -1,0 +1,2 @@
+export * from './UserController';
+export * from './RefreshToken';
